+++ src/by/it/okatov/jd02_02/Buyer.java
@@ -3,174 +3,146 @@ package by.it.okatov.jd02_02;
 import java.util.HashMap;
 import java.util.Map;
 
-class Buyer extends Thread implements IBuyer, IUseBacket {
-    private static boolean isElder;
-    private static final float ELDERY_COEF = 1.5f;
-    private static int numberOfBuyer;
-    private final Map<String, Integer> tmp = new HashMap<>();//Список покупок
-    private static int sum = 0;
+class Buyer extends Thread implements IBuyer, IUseCart {
+    Cart cart;
+    private static boolean elderMark;
+    private static final float ELDER_COEF = 1.5f;
+    static volatile int globalCount = 0;
 
-    public int getSum() {
-        return sum;
+    public static boolean isElderMark() {
+        return elderMark;
     }
 
-    public void setSum(int pm) {
-        sum = pm;
+    public static void setElderMark(boolean elderMark) {
+        Buyer.elderMark = elderMark;
     }
 
+    private final Map<String, Integer> goods = new HashMap<>();
 
-    boolean isElder() {
-        return isElder;
+    public Map<String, Integer> getGoods() {
+        return goods;
     }
 
-    static void setEldery(boolean eldery) {
-        isElder = eldery;
+    public void setGoods(Map.Entry<String, Integer> entry) {
+        this.goods.put(entry.getKey(), entry.getValue());
     }
 
-    Buyer(int num) {
-        super(String.format("Buyer #%d ", num));
-        numberOfBuyer = num;
-        Manager.addBuyer();
+    public Buyer(int number) {
+        super("Buyer №" + number);
+        Manager.buyerAddToShop();
     }
 
-    Buyer(int num, boolean isElder) {
-        super(String.format("Buyer (pensioneer) #%d ", num));
-        setEldery(true);
-        numberOfBuyer = num;
-        Manager.addBuyer();
+    Buyer(int number, boolean elderMark) {
+        super("Buyer (Pensioneer) №" + number);
+        setElderMark(elderMark);
+        Manager.buyerAddToShop();
     }
-
 
     @Override
     public void run() {
-        enterToMarket(); //Войти в магазин
-        takeCart(); //Взять тележку
-        chooseGoods(); //Выбрать товары
-        putGoodsToCart(); //Положить товары в корзину
+        enterToMarket();
+        takeCart();
+        chooseGoods();
+        putGoodsToCart();
         goToQueue();
-        goOut();//Уйти из магазина (разумеется, не заплатив ХDDD)
+        returnCart();
+        goOut();
     }
 
     @Override
     public void enterToMarket() {
-        //Utils.threads.add(this);
-        System.out.println(this + "enters the supermarket");
-        Utils.GLOBAL_COUNTER++;//Увеличиваем счетчик покупателей
+        System.out.println(this + " enter to shop");
+        globalCount++;
+    }
+
+    @Override
+    public void takeCart() {
+        int timeout = Utils.getRandom(0, 2);
+        System.out.println(this + " takes cart");
+        if (isElderMark()) {
+            Utils.waitForSeconds(timeout * ELDER_COEF);
+        } else {
+            Utils.waitForSeconds(timeout);
+        }
+        cart = new Cart();
     }
 
     @Override
     public void chooseGoods() {
-        System.out.println(this + "starts to collect goods");
-        int timeout = Utils.getRandom(1, 3);
-        if (isElder()) {//Если пенсионер
-            Utils.waitForSeconds(timeout * ELDERY_COEF);
+        int timeout = Utils.getRandom(0, 2);
+        System.out.println(this + " started to choose goods");
+        if (isElderMark()) {
+            Utils.waitForSeconds(timeout * ELDER_COEF);
         } else {
             Utils.waitForSeconds(timeout);
         }
 
+        System.out.println(this + " finished to choose goods");
+    }
 
-        System.out.println(this + "finishes collecting goods");
-
+    @Override
+    public void putGoodsToCart() {
+        System.out.println(this + " puts goods to cart: ");
+        //Map<String, Integer> tmp = new HashMap<>();
+        int index = 0;
+        String key;
+        int val;
+        for (Map.Entry<String, Integer> entry : Utils.gethMapOfGoods().entrySet()) {
+            if (Utils.getRandom(0, 4) == index) {
+                continue;
+            } else if (index == 4) {
+                break;
+            }
+            key = entry.getKey();
+            val = entry.getValue();
+            System.out.println(String.format("%s решил взять %s за %d р.", this, key, val));
+            goods.put(key, val);
+            index++;
+        }
     }
 
     @Override
     public void goToQueue() {
         synchronized (this) {
-            BuyersQueue.add(this);
+            if (isElderMark()) {
+                QueueBuyers.addElder(this);
+            } else {
+                QueueBuyers.add(this);
+
+            }
             try {
                 System.out.println(this + " added to queue");
-                wait();
-                System.out.println(this + " leaves queue");
+                wait();///------------------------------------------------------------------------------------
+                System.out.println(this + " leaves the queue");
             } catch (InterruptedException e) {
                 throw new RuntimeException(e);
             }
         }
 
+
+    }
+
+    @Override
+    public void returnCart() {
+        int timeout = Utils.getRandom(0, 2);
+        System.out.println(this + " returns cart");
+        if (isElderMark()) {
+            Utils.waitForSeconds(timeout * ELDER_COEF);
+        } else {
+            Utils.waitForSeconds(timeout);
+        }
+        cart.clearCart();
     }
 
     @Override
     public void goOut() {
-
-        Utils.GLOBAL_COUNTER--;//Уменьшаем количество покупателей
-        System.out.println(this + "returns a cart in the supermarket");
-        if (isElder()) {//Было решено, что на выход из магазина тоже нужно время.
-            Utils.waitForSeconds(1 * ELDERY_COEF);
-        } else {
-            Utils.waitForSeconds(1);
-        }
-        //System.out.println(printCheck());
-        System.out.println(this + "leaves the supermarket");
-        Manager.removeBuyer();
-        //Utils.threads.remove(this);
+        System.out.println(this + " leaves the supermarket");
+        Manager.buyerLeaveShop();
+        globalCount--;
     }
-
-    @Override
-    public String printCheck(int spaces) {
-        int payment = 0;
-        StringBuilder sb = new StringBuilder();
-        //setDelimiters
-        StringBuilder delimiter = Utils.setDelimiters(spaces);
-        delimiter.toString();
-
-        sb.append(delimiter).append("Список покупок ").append(this).append(":").append("\n");
-        sb.append(delimiter).append("┌──────────────────────────────────────┐\n");
-
-
-        for (Map.Entry<String, Integer> entry : tmp.entrySet()) {
-            payment += entry.getValue();
-            setSum(payment);
-            sb.append(delimiter).append("│");
-            sb.append(entry.getKey()).append("..................")
-                    .append(entry.getValue()).append(" р.│\n");
-        }
-
-        sb.append(delimiter).append("└──────────────────────────────────────┘\n");
-        sb.append(delimiter).append("Сумма счета ").append(this).append(": ").append(getSum()).append("\n");
-        return sb.toString();
-    }
-
-
-    @Override
-    public void takeCart() {
-        int timeout = Utils.getRandom(1, 3);
-        if (isElder()) {
-            Utils.waitForSeconds(timeout * ELDERY_COEF);
-        } else {
-            Utils.waitForSeconds(timeout);
-        }
-        System.out.println(this + "takes a cart in the supermarket");
-    }
-
-
-    @Override
-    public void putGoodsToCart() {
-        int goodsQuant = Utils.getRandom(0, 4);
-        int i = 0;
-        int timeout = Utils.getRandom(1, 3);
-
-        for (Map.Entry<String, Integer> entry : Utils.gethMap().entrySet()) {
-
-            if (isElder()) {
-                Utils.waitForSeconds(timeout * ELDERY_COEF);
-            } else {
-                Utils.waitForSeconds(timeout);
-            }
-            if (i >= goodsQuant) {
-                break;
-            }
-            String key = entry.getKey();
-            int val = entry.getValue();
-            tmp.put(key, val);
-            System.out.println(this + "puts " + key + " for $" + val);
-            i++;
-        }
-    }
-
 
     @Override
     public String toString() {
         return getName();
     }
-
-
 }
