+++ src/by/it/okatov/jd02_02/Chart.txt
@@ -1,0 +1,67 @@
+0 0
+1 0
+2 2
+3 6
+4 8
+5 8
+6 8
+7 8
+8 11
+9 9
+10 8
+11 10
+12 11
+13 13
+14 14
+15 12
+16 12
+17 14
+18 18
+19 17
+20 15
+21 20
+22 17
+23 18
+24 16
+25 16
+26 13
+27 13
+28 12
+29 12
+30 9
+31 9
+32 9
+33 7
+34 5
+35 7
+36 8
+37 9
+38 9
+39 12
+40 11
+41 13
+42 14
+43 14
+44 17
+45 17
+46 15
+47 19
+48 21
+49 21
+50 19
+51 16
+52 18
+53 15
+54 14
+55 14
+56 13
+57 12
+58 12
+59 10
+60 9
+61 9
+62 8
+63 8
+64 5
+65 3
+66 8
