--- src/by/it/okatov/jd02_02/IUseBacket.java
@@ -1,7 +1,0 @@
-package by.it.okatov.jd02_02;
-
-interface IUseBacket {
-    void takeCart(); //взял корзину
-
-    void putGoodsToCart(); //положил выбранный товар в корзину
-}
