+++ src/by/it/okatov/jd02_02/IUseCart.java
@@ -1,4 +1,4 @@
-package by.it.okatov.jd02_02.Test;
+package by.it.okatov.jd02_02;
 
 public interface IUseCart {
     void takeCart(); //взял корзину
