+++ src/by/it/okatov/jd02_02/QueueBuyers.java
@@ -1,4 +1,4 @@
-package by.it.okatov.jd02_02.Test;
+package by.it.okatov.jd02_02;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
